requirements.txt
torch
torchvision
opencv-python
numpy
matplotlib
seaborn
pillow
scikit-learn
scikit-image
tqdm